move switches to settings
